shift link and skill display to utils
